KOGIS_PRJ-1382: adaption of README.md | added secrets.yml | review suggestions ltshb
* added the secrets.yml in ./scripts so that summon can inject gopass secrets into env vars
* added description in README.md
* adapted scripts as suggested in PR review by ltshb
