+++ .gitignore
@@ -71,6 +71,7 @@ nose2-junit.xml
 .volumes
 **/logs
 /scripts/xxxxxxl_asset_file.zip
+/scripts/dummy_asset_for_multipart_testing.zip
 
 # Node modules
 node_modules/*
