+++ scripts/secrets.yml
@@ -1,0 +1,13 @@
+common: {}
+
+dev:
+  STAC_USER: !var infra-gopass-bgdi/service-stac.dev.bgdi.ch/kogis_admin username
+  STAC_PASSWORD: !var infra-gopass-bgdi/service-stac.dev.bgdi.ch/kogis_admin password
+
+int:
+  USERNAME_INT: !var infra-gopass-bgdi/service-stac.int.bgdi.ch/kogis_admin username
+  PASSWORD_INT: !var infra-gopass-bgdi/service-stac.int.bgdi.ch/kogis_admin password
+
+prod:
+  USERNAME_PROD: !var infra-gopass-bgdi/service-stac.prod.bgdi.ch/kogis_admin username
+  PASSWORD_PROD: !var infra-gopass-bgdi/service-stac.prod.bgdi.ch/kogis_admin password
